Feature: select the AES-192 cipher for the P-384 curve by default
The SEC 1 recommends to select the derived key length as a half of the curve strength.
That recommendation is called to protect the key security from being compromised.

However, the umbracle/ecies broke that recommendation by chosing a stronger cipher.

So, here we restore the default to recommended values.
Also, confirmed that the go-ethereum ECIES implementation uses the same values for P-384.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>